backend/ai_providers/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
httpx
psycopg2-binary
python-jose
google-generativeai